move symbols out of CodegenLayout.Companion
